// src/app/core/Models/IAccount.ts
export interface IAccount {
  claim: any
  hasAzureADAccount: boolean
  createdDate: string
  updatedDate: any
  id: number
  fullName: string
  phoneNumber: string
  email: string
  emiratesIdNumber: string
  arabicSurname: string
  englishSurname: string
  jobTitle: string
  gender: number
  birthDate: string
  emiratesIdPath: string
  employeeIdNumber: string
  permissionToEnterScore: boolean
  status: string
  nationalityId: number
  relativeRelationId: number
  password: any
  nickName: string
  isActive: boolean
  scope: string
  roles: any
  }
  
  